requirements.txt
fastapi[all]==0.101.0
uvicorn[standard]==0.23.2
httpx==0.24.1
# MongoDB
motor==3.2.0

pydantic==2.1.1
bcrypt==4.0.1

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pytest==7.4.0
pytest-cov==4.1.0
mock==5.1.0
pytest-httpx==0.23.1

Faker==18.11.2

aioboto3==11.2.0
pandas==2.0.3

colorlog==6.7.0

rollbar==0.16.3
xlrd==2.0.1

# Queue
celery==5.3.4
redis==5.0.1

# google
google-api-core==2.15.0
google-api-python-client==2.111.0
google-auth==2.25.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
